Core, Graph: Moved GraphFeatures to .graph plug-in.
