+++ plugins/org.eclipse.elk.core.meta/META-INF/MANIFEST.MF
@@ -16,8 +16,7 @@ Require-Bundle: org.eclipse.xtext;bundle-version="2.9.0",
  org.eclipse.xtext.util,
  org.eclipse.xtend.lib,
  org.eclipse.emf.common,
- org.eclipse.elk.graph,
- org.eclipse.elk.core
+ org.eclipse.elk.graph
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.elk.core.meta,
  org.eclipse.elk.core.meta.metaData,
