+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/data/LayoutAlgorithmData.java
@@ -15,9 +15,9 @@ import java.util.Map;
 import java.util.Set;
 
 import org.eclipse.elk.core.AbstractLayoutProvider;
-import org.eclipse.elk.core.options.GraphFeature;
 import org.eclipse.elk.core.util.IFactory;
 import org.eclipse.elk.core.util.InstancePool;
+import org.eclipse.elk.graph.properties.GraphFeature;
 
 import com.google.common.collect.Maps;
